refactor: move LockKeyProvider to application.common
