+++ src/main/kotlin/com/dobby/backend/application/common/LockKeyProvider.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.application.aop
+package com.dobby.backend.application.common
 
 interface LockKeyProvider {
     fun getLockKey(): String
